move util files for view
